requirements/fb.txt
django-facebook-graph>=0.2.0